Create logging module
